Simple protection against non-existing data dir
In my case it threw "Unable to write the object file! Please check if PHP has write permission.", but the permissions were there. I figured it must be missing the data dir and I didn't find anything about creating one manually in the doc so... Not sure how it worked for other people... maybe it's my setup (PHP 7.2.19-0ubuntu0.18.04.2) but, these couple of lines helped.